fix(irc): construct runner before inline factories; re-add hub after a bridge claims a transport
Two more PR #7401 codex follow-ups (not multi-main-registry):

- r3760356006 (sdk.ts): the extension runner (the startup-abort shutdown owner)
  sat below the inline-factory loop, so a throw from a later inline factory -
  after an earlier one already claimed an IRC namespace - hit the !hasSession
  catch with no runner: emitSessionShutdownEvent no-op'd and the earlier claim's
  transport / proxies leaked. Move the runner construction + credentialDisabledTarget
  assignment above the inline loop (and provider registration). hasHandlers/emit
  read extensionsResult.extensions live, so inline extensions the loop pushes are
  still seen at shutdown.

- r3760355996 (sdk.ts / hub): createTools() builds the built-in slate before
  extensions load, gating hub on isIrcEnabled -> hasRemoteTransport(). A bridge
  installs its transport during load, so a leaf root (task.maxRecursionDepth=0,
  no spawn-based peers) dropped hub for the whole session while the prompt still
  advertised the remote peers. Re-add the hub built-in after extension load if a
  transport was claimed and it is not already present (mirrors the goal/write
  late-registration precedent).

Tests (test/sdk-irc-extension.test.ts, renamed from sdk-startup-abort-irc): an
inline factory that throws after an earlier IRC claim releases that claim; a
leaf-root bridge session gets hub active. Both verified to fail without their fix.

murmur-nseb / murmur-up8k (session-based registries).
